[bundler] Move babel configuration out into it's own preset and a package -- babel-preset-startupjs. Deprecate usage of old config directly from bundler. [templates][styleguide] Update to use new babel config.
