+++ packages/babel-preset-startupjs/.gitignore
@@ -1,0 +1,8 @@
+# OSX
+.DS_Store
+
+node_modules/
+npm-debug.log
+yarn-error.log
+
+/lib/
