+++ packages/babel-preset-startupjs/README.md
@@ -1,0 +1,22 @@
+# babel-preset-startupjs
+> Babel preset for compiling StartupJS app on server, web, native
+
+## Usage in StartupJS projects
+
+StartupJS projects include this configuration by default. **You don't need to install it separately in StartupJS projects.**
+
+## Usage outside of StartupJS
+
+If you want to use this preset in a custom project, just plug it in `.babelrc`:
+
+```json
+{
+  presets: ['startupjs']
+}
+```
+
+## Licence
+
+MIT
+
+(c) Decision Mapper - http://decisionmapper.com
