+++ packages/babel-preset-startupjs/constants.json
@@ -1,0 +1,3 @@
+{
+  "LOCAL_IDENT_NAME": "[local]_[hash:base64:5]_"
+}
