+++ packages/bundler/package.json
@@ -15,42 +15,30 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-proposal-decorators": "^7.7.0",
-    "@babel/plugin-transform-runtime": "^7.6.2",
     "@babel/polyfill": "^7.7.0",
-    "@babel/runtime": "^7.7.2",
     "@dmapper/stylus-hash-plugin": "0.0.1",
     "@mdx-js/loader": "^1.5.8",
     "@mdx-js/mdx": "^1.5.8",
-    "@startupjs/babel-plugin-dotenv-import": "1002.0.0",
-    "@startupjs/babel-plugin-react-css-modules": "1701.0.0",
     "@startupjs/postcss-rem-to-pixel": "1004.1.2",
     "@startupjs/react-native-stylus-transformer": "1201.2.0",
     "@svgr/webpack": "5.2.0",
     "assets-webpack-plugin": "^3.9.10",
     "autoprefixer": "^9.6.0",
     "babel-loader": "^8.0.6",
-    "babel-plugin-module-resolver": "^3.2.0",
-    "babel-plugin-react-native-dynamic-stylename-to-style": "^0.13.2",
-    "babel-plugin-react-native-platform-specific-extensions": "^1.1.1",
-    "babel-plugin-react-native-web-pass-classname": "^0.2.1",
-    "babel-plugin-react-pug-classnames": "^0.3.1",
-    "babel-plugin-transform-react-pug": "^7.0.1",
+    "babel-preset-startupjs": "^0.19.2",
     "css-loader": "^1.0.0",
     "file-loader": "^2.0.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "generic-names": "^1.0.3",
     "lodash": "4.x",
     "metro-config": "^0.57.0",
-    "metro-react-native-babel-preset": "^0.57.0",
     "mini-css-extract-plugin": "^0.8.0",
     "moment": "^2.0.0",
     "moment-locales-webpack-plugin": "^1.0.7",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "postcss-loader": "^3.0.0",
     "progress-bar-webpack-plugin": "^2.1.0",
-    "react-hot-loader": "~4.12.13",
+    "react-hot-loader": "^4.12.13",
     "react-native-css-transformer": "^1.2.4",
     "react-native-svg-transformer": "0.14.3",
     "style-loader": "^0.21.0",
