+++ packages/bundler/webpack.web.config.js
@@ -9,7 +9,7 @@ const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin')
 const TerserPlugin = require('terser-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
-const { LOCAL_IDENT_NAME } = require('./buildOptions')
+const { LOCAL_IDENT_NAME } = require('babel-preset-startupjs/constants')
 const { getJsxRule } = require('./helpers')
 const autoprefixer = require('autoprefixer')
 const rem2pixel = require('@startupjs/postcss-rem-to-pixel')
