+++ packages/startupjs/babel.js
@@ -1,0 +1,1 @@
+module.exports = require('babel-preset-startupjs')
