+++ packages/startupjs/package.json
@@ -14,6 +14,7 @@
     "@startupjs/init": "^0.19.2",
     "@startupjs/orm": "^0.19.2",
     "@startupjs/react-sharedb": "^0.19.2",
-    "@startupjs/server": "^0.19.2"
+    "@startupjs/server": "^0.19.2",
+    "babel-preset-startupjs": "^0.19.2"
   }
 }
