+++ styleguide/babel.config.js
@@ -1,10 +1,8 @@
-const getConfig = require('startupjs/bundler').babelConfig
-
-module.exports = function (api) {
-  const config = getConfig(api, {
-    legacyClassnames: false,
-    alias: {}
-  })
-
-  return config
+module.exports = {
+  presets: [
+    ['startupjs/babel', {
+      legacyClassnames: false,
+      alias: {}
+    }]
+  ]
 }
